requirements.txt
asgiref==3.2.7
attrs==19.3.0
backcall==0.1.0
bleach==3.1.5
bson==0.5.8
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
Django==2.2.13
django-cors-headers==3.3.0
djangorestframework==3.11.0
djongo==1.2.38
dnspython==1.16.0
entrypoints==0.3
google-api-core==1.20.1
google-api-python-client==1.9.3
google-auth==1.17.2
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-speech==1.3.2
googleapis-common-protos==1.52.0
grpcio==1.29.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
json5==0.9.4
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.1.2
jupyterlab-server==1.1.3
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
packaging==20.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pymongo==3.10.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
requests==2.23.0
rsa==4.6
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.2.4
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
